Use usize to support long input (#72)
* Use usize to support long input

* Remove unnecessary condition

* fix

* Unwrap results

* Update README.md

* Remove unnecessary comments

* fmt

+++ vibrato/src/dictionary/lexicon/map.rs
@@ -33,7 +33,7 @@ impl WordMap {
     pub fn common_prefix_iterator<'a>(
         &'a self,
         input: &'a [char],
-    ) -> impl Iterator<Item = (u32, u16)> + 'a {
+    ) -> impl Iterator<Item = (u32, usize)> + 'a {
         self.trie.common_prefix_iterator(input).flat_map(move |e| {
             self.postings
                 .ids(usize::from_u32(e.value))
@@ -45,7 +45,7 @@ impl WordMap {
     pub unsafe fn common_prefix_iterator_unchecked<'a>(
         &'a self,
         input: &'a [char],
-    ) -> impl Iterator<Item = (u32, u16)> + 'a {
+    ) -> impl Iterator<Item = (u32, usize)> + 'a {
         self.trie.common_prefix_iterator(input).flat_map(move |e| {
             self.postings
                 .ids_unchecked(usize::from_u32(e.value))
